MS5/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
pika

PyJWT
grpcio
grpcio-tools
protobuf
google-api-python-client
dotenv